+++ tests/Community/Training/Easy/DistributingCandy/CGTest.php
@@ -1,0 +1,88 @@
+<?php
+
+declare(strict_types=1);
+
+namespace CyrilVerloop\Codingame\Tests\Community\Training\Easy\DistributingCandy;
+
+use CyrilVerloop\Codingame\Tests\PuzzleTest;
+use CyrilVerloop\Codingame\Community\Training\Easy\DistributingCandy\DistributingCandy;
+
+/**
+ * Tests for the "Distributing candy" puzzle.
+ *
+ * @covers \CyrilVerloop\Codingame\Community\Training\Easy\DistributingCandy\DistributingCandy
+ * @group distributingCandy
+ * @medium
+ */
+final class CGTest extends PuzzleTest
+{
+    public function setUp(): void
+    {
+        $this->puzzle = new DistributingCandy();
+    }
+
+    /**
+     * Test that the code can be executed for "Test 1".
+     *
+     * @group distributingCandy_test1
+     */
+    public function testCanExecuteTest1(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/01 - test 1.txt',
+            2 . PHP_EOL
+        );
+    }
+
+    /**
+     * Test that the code can be executed for "Test 2".
+     *
+     * @group distributingCandy_test2
+     */
+    public function testCanExecuteTest2(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/02 - test 2.txt',
+            5 . PHP_EOL
+        );
+    }
+
+    /**
+     * Test that the code can be executed for "Test 3".
+     *
+     * @group distributingCandy_test3
+     */
+    public function testCanExecuteTest3(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/03 - test 3.txt',
+            23 . PHP_EOL
+        );
+    }
+
+    /**
+     * Test that the code can be executed for "Test 4".
+     *
+     * @group distributingCandy_test4
+     */
+    public function testCanExecuteTest4(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/04 - test 4.txt',
+            13 . PHP_EOL
+        );
+    }
+
+    /**
+     * Test that the code can be executed for "Test 5".
+     *
+     * @group distributingCandy_test5
+     */
+    public function testCanExecuteTest5(): void
+    {
+        $this->expectExecuteOutputAnswer(
+            __DIR__ . '/input/05 - test 5.txt',
+            20 . PHP_EOL
+        );
+    }
+}