+++ CHANGELOG.md
@@ -35,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "A mountain of a mole hill".
 - Tests for "Where's Wally".
 - Tests for "Feature extraction".
+- Tests for "Distributing candy".
 
 ## [3.7.0] - 2022-03-31
 ### Added